Replace ML-based spam detection with manual review
Unfortunately the machine-learning based approach for detecting new
referrer spam domain names in the analytics data of Google Analytics
accounts did not work reliably enough for websites with different usage
patterns. So for now I replaced the automatic spam-detection with a
simple manual review process.

The `find-spam-domains` action will now download the last 90 days of
analytics data from the specified Google Analytics account and list
a referrer domain names which are currently NOT recognized as spam.
If you find a domain name which is spam, you uncomment the line.
When you are finished with the review, you save the file and close the
editor. All domain names which were uncommented are added to your
personal referrer spam domain list and henceforce combined with the
list of known referrer spam domain names that are provided by the
community.

Changes:

- Remove the machine learning spam detection. I might try it again in
  the future. But for now the community-maintained referrer spam lists
  will have to suffice
- Introduce a personal repository spam domain repository
- Use a separate config directory for the domain lists
- Use Windows line endings on Windows
- Open notepad on Windows (instead of Vim) for editing the domain list
- Open the Google Analytics oAuth authorizatin URL with the users default
  browser using github.com/skratchdot/open-golang
- Include file paths in help text
- Updated the roadmap