Base protection level on domain names, not filters
The spam-protection status calculation is now based on the number of
spam domain names that are blocked by the existing filters - opposed to
the number of filters which are up-to-date.

The previous approach gave a false impression of the spam protection
status. The displayed percentage was usually to low. In the worst case
many filters were outdated when the new domain names had all a unique
first characters - because the filter segments are based on the first
character of the spam domain names.